+++ DelvUI/changelog.md
@@ -4,7 +4,7 @@
     + The tags can be added to the label by clicking on them.
     + Hovering with the cursor on top of a tag shows an example of what they do.
 
-- Added option to set Dead Backdrop color for unit- and party frames.
+- Added option to set a Death Indicator Background Color for unit- and party frames. Disabled by default.
 - Added Buff/Debuffs for the Focus Target
 - Added a sub-option to Hide DelvUI outside of combat to always show in duties.
 - Re-enabled labels for chunked bars.